If deletion fails with a `WriteException` support exclusion creation via error dialog
